requirements.txt
flask
pandas
xgboost
scikit-learn
joblib
flask_cors
flask-bcrypt
Flask-Mail
python-dotenv